[fix] BBox int to double

+++ src/globals.h
@@ -8,10 +8,10 @@
 typedef std::vector<std::vector<std::vector<std::unordered_map<std::string, double>>>> pointCollection;
 
 typedef struct {
-    int left;   // min longitude    [-180, 180[
-    int top;    // min latitude     [ -90,  90[
-    int right;  // max longitude    [-180, 180[
-    int bottom; // max latitude     [ -90,  90[
+    double left;   // min longitude    [-180, 180[
+    double top;    // min latitude     [ -90,  90[
+    double right;  // max longitude    [-180, 180[
+    double bottom; // max latitude     [ -90,  90[
 } BBox;
 
 #endif //QUALITAIR_GLOBALS_H
